execute_function: generic function definition
Define a new generic function called "execute_function" which
takes function_name & args (which can be a marshalled JSON string)
and returns the result of executed function string (this can also be
a marshalled JSON string).

The execute_function is the host function implemented using the extism
sdk which runs the guest wasm binary.