compute-decorator: extism decorator function to invoke host-function (#2)
Define "compute" wrapper function which internally invokes host function
called "host_callback". The inner closure function takes the function
name to which compute is wrapper and its payload to build the input
args for the host_callback function.

The host_callback function takes following arguments:
  * type: same as the decorator function name
  * payload: generated from the function name to which compute is
    decorator and its arguments. We assume that there is only one
    argument for now.

We make an assumption here that host_callback is the only host function
defined as part of the script, that is why it is safe to assume the
index to be "0" when invoking the host function using __invoke_host_func
implementation.