signal-support: Add signal decorator to send signal data
Add another decorator called wait_for_signal which can be used to wait
until the signal data is returned by the host_callback function back to
the guest.

As part of the wait_for_signal payload we don't send any data, but
just the name of signal function for which it is waiting on.