query: Add support for query
Define a new decorator query, that stores the query functions in a list.
The function is invoked when the query is received as part of the host
function invocation.

We update the host function invocation functionality to support the
query workflow. As part of host function execution, it can be the case,
that function is invoked with compute or signal type event, but response
is for the query. In such cases, we execute the query to get the result
and make another host function invocation with the query result. Later,
the host will send the response with non-query type event and the host
function invocation loop will break.