update-operation: Support update operation for guest (#9)
Add ability to receive update operation from the host and run the update
function exported by the guest. Return the result of the function back
to the host. If the guest was already waiting on signal condition, then
set "waiting_on_condition" to in the response payload of the update
operation.